// src/interfaces/ISingleMovie.ts
export default interface ISingleMovie {
  adult: boolean
  backdrop_path: string | null
  belongs_to_collection: {
    id: number
    name: string
    poster_path: string
    backdrop_path: string
  } | null
  budget: number
  genres: [
    {
      id: number
      name: string
    }
  ]
  homepage: string | null
  id: number
  imdb_id: string | null
  original_language: string
  original_title: string
  overview: string | null
  popularity: number
  poster_path: string | null
  production_companies: [
    {
      id: number
      logo_path: string
      name: string
      origin_country: string
    }
  ]
  production_countries: [
    {
      iso_3166_1: string
      name: string
    }
  ]
  release_date: string
  revenue: number
  runtime: number | null
  spoken_languages: [
    {
      english_name: string
      iso_639_1: string
      name: string
    }
  ]
  status:
    | 'Rumored'
    | 'Planned'
    | 'In Production'
    | 'Post Production'
    | 'Released'
    | 'Canceled'
  tagline: string | null
  title: string
  video: boolean
  vote_average: number
  vote_count: number
}